+++ internal/data/category_aggregation_store.go
@@ -6,7 +6,7 @@ import (
 )
 
 type CategoryAggregationStore interface {
-	AggregateCategory(cat models.Category, date datatypes.Date, elapsedTime int) (*models.CategoryAggregation, error)
+	AggregateCategory(cat models.Category, date datatypes.Date, elapsedTime int64) (*models.CategoryAggregation, error)
 	GetCategoryAggregation(categoryId models.CategoryId, date datatypes.Date) (*models.CategoryAggregation, bool)
 	GetCategoryAggregations() ([]*models.CategoryAggregation, error)
 	GetCategoryAggregationsByDate(date datatypes.Date) ([]*models.CategoryAggregation, error)
